Fixes, sprite color change; more debug visuals
I reviewed a lot of my code, and found that some baselining values were
not changing the way they should have been.  This has been fixed.

I also made it so the sprite is only going to turn red in case of error;
this reduces the number of factors for a subject to have to worry about.