Bugfixes
Fixed:
-When recording, should no longer cause round 3 and 4 to fuse together,
and then crash.
-The glider should no longer be able to rise when the lownoise threshold
is passed, points should not be gainable.
-First baselining session now actually informs the second thresholds.

Bandaid solution:  There is a lag between the NeuroTrain timer, and the
timer in the game.  I'm pretty sure the NeuroTrain one is the faulty
one, but I've added 8 seconds to the baseline fixation display as a
bandaid solution to one period ending before the other.